fix(core): improves variable scope within IfStatements
Knowing where to assign variables within a program can be an odd challenge. CoffeeScript normally
places all implicitly defined variables in a list of declarations at the top scope of a function,
which makes it a bit difficult to debug where the variable is originally defined. This update should
more efficiently place initially defined variables in the correct scope of functions and
IfStatements (still room for improvement most likely).